python-agents/orderagent/requirements.txt
langchain==0.3.26  # 最新稳定版
langchain-community==0.3.26
langchain-core==0.3.67
langchain-openai==0.3.9  # 满足langchain-deepseek的要求
langchain-deepseek==0.1.3
langsmith==0.4.4
langgraph==0.5.0
fastapi
uvicorn
pydantic
python-dotenv